pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyscripts"
version = "1.0.0"
dependencies = [
	"libretranslatepy",
	"matplotlib",
	"pillow",
	"requests",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
